pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "python-cmr"
version = "0.11.0-a0"
description = "Python wrapper to the NASA Common Metadata Repository (CMR) API."
authors = ["python_cmr <nasa/python_cmr@github.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/nasa/python_cmr"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Framework :: IPython",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    # "Typing :: Typed",
]
packages = [{ include = "cmr" }]

[tool.poetry.dependencies]
python = "^3.8.1"
requests = "^2.26.0"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
flake8 = "^7.0.0"
pytest = "^6.2.5"
vcrpy = "^5.1.0"
